tdg-probe-social/requirements.txt
Flask_Migrate==2.0.4
oauth2==1.9.0.post1
flask_marshmallow==0.8.0
Flask_Script==2.0.5
Flask_SQLAlchemy==2.2
requests==2.18.1
alembic==0.9.3
aiohttp==2.2.3
SQLAlchemy==1.1.11
Flask==0.12.2
python_dateutil==2.6.1
